ProyectoFinalBI/I Parte - twitterstreaming/requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
kafka-python==1.4.7
oauthlib==3.1.0
psutil==5.6.7
py4j==0.10.7
PySocks==1.7.1
pyspark==2.4.4
pyspark-stubs==2.4.0.post7
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
tweepy==3.8.0
urllib3==1.25.8
